requirements_phase1_extended.txt
# Speed Climbing Performance Analysis - Phase 1 Extended Dependencies
# Updated: 2025-11-12
# Python: 3.11+

# ========================================
# Core Computer Vision
# ========================================
opencv-python==4.8.1.78
mediapipe==0.10.8
scikit-image==0.22.0          # Advanced image processing

# ========================================
# Numerical Computing
# ========================================
numpy==1.24.3
scipy==1.11.4
pandas==2.1.4

# ========================================
# Machine Learning
# ========================================
torch==2.1.2
torchvision==0.16.2
scikit-learn==1.3.2

# ========================================
# Fuzzy Logic
# ========================================
scikit-fuzzy==0.4.2

# ========================================
# Visualization
# ========================================
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# ========================================
# Audio/Video Processing (NEW for Phase 1 Extended)
# ========================================
yt-dlp==2024.8.6              # YouTube downloader
pydub==0.25.1                 # Audio manipulation
librosa==0.10.1               # Audio analysis (for beep detection)
ffmpeg-python==0.2.0          # FFmpeg wrapper
soundfile==0.12.1             # Audio file I/O (required by librosa)

# ========================================
# Tracking & Filtering (NEW)
# ========================================
filterpy==1.4.5               # Kalman filter for tracking

# ========================================
# Utilities
# ========================================
pyyaml==6.0.1
tqdm==4.66.1
requests==2.31.0              # For API calls
pillow==10.1.0                # Image processing utilities

# ========================================
# Testing
# ========================================
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0           # Mocking for tests

# ========================================
# Development Tools (optional)
# ========================================
jupyter==1.0.0
ipywidgets==8.1.1
black==23.12.0                # Code formatting
flake8==6.1.0                 # Linting
mypy==1.7.1                   # Type checking

# ========================================
# Google Colab (optional)
# ========================================
# google-colab==1.0.0         # Uncomment if using Colab

# ========================================
# NOTES:
# ========================================
# - ffmpeg must be installed separately on Windows:
#   Download from: https://www.gyan.dev/ffmpeg/builds/
#   Add to PATH: C:\ffmpeg\bin
#
# - For Intel ARC NPU optimization:
#   pip install intel-extension-for-pytorch==2.1.10
#
# - To install all at once:
#   pip install -r requirements_phase1_extended.txt
#
# - To upgrade existing:
#   pip install --upgrade -r requirements_phase1_extended.txt
